requirements.txt
numpy>=1.19.0
scikit-learn>=0.24.0
joblib>=1.0.0
pyyaml>=5.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
pandas>=1.3.0
aiogram>=3.14.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
openpyxl>=3.0.10